Move model_validator.py to validations/semantic_manifest_validator.py.
